Hadron.Basic

Description

Low level building blocks for working with Hadoop streaming.

We define all the base types for MapReduce and export map/reduce maker functions that know how to deal with ByteString input and output.

Types

type Key = ByteString

type CompositeKey = [Key]

type Mapper a k b = Conduit a (ResourceT IO) (k, b)

A Mapper parses and converts the unbounded incoming stream of input data into a stream of (key, value) pairs.

A mapper is responsible for signaling its own Nothing to signify it is done writing. Hadron will not automatically mark EOF on the OutputStream, which is basically stdout.

type Reducer k a r = Conduit (k, a) (ResourceT IO) r

A reducer takes an incoming stream of (key, value) pairs and emits zero or more output objects of type r.

Note that this framework guarantees your reducer function (i.e. the conduit you supply here) will see ONLY keys that are deemed equivalent based on the MROptions you supply. Different keys will be given to individual and isolated invocations of your reducer function. This is a very central abstraction (and one of the few major ones) provided by this framework.

It does not matter if you supply your own EOF signals via Nothing as Hadron will simply discard them before relaying over to stdout and supply its own EOF based on when its input is finished.

Hadoop Utilities

emitCounter

Arguments

:: ByteString
Group name

-> ByteString
Counter name

-> Integer
Increment

-> IO ()

Emit a counter to be captured, added up and reported by Hadoop.

hsEmitCounter :: ByteString -> Integer -> IO ()

Emit counter from this library's group

emitStatus :: ByteString -> IO ()

Emit a status line.

getFileName :: MonadIO m => m FilePath

Get the current filename from Hadoop ENV. Useful when writing Mappers and you would like to know what file you're currently dealing with.

MapReduce Construction

mapReduceMain

Arguments

:: MROptions

-> Prism' ByteString a
Serialization function for the in-between data a.

-> Mapper ByteString CompositeKey a

-> Reducer CompositeKey a ByteString
Reducer for a stream of values belonging to the same key.

-> IO ()

A default main that will respond to map and reduce commands to run the right phase appropriately.

This is the recommended main entry point to a map-reduce program. The resulting program will respond as:

./myProgram map
./myProgram reduce

mapReduce

Arguments

:: MROptions

-> Prism' ByteString a
Serialization for data between map and reduce stages

-> Mapper ByteString CompositeKey a

-> Reducer CompositeKey a ByteString

-> (IO (), IO ())

data MROptions

Options for a single-step MR job.

Constructors

MROptions

Fields

_mroPart :: PartitionStrategy

Number of segments to expect in incoming keys. Affects both hadron program's understanding of key AND Hadoop's distribution of map output to reducers.

_mroComparator :: Comparator

_mroNumMap :: Maybe Int

Number of map tasks; Nothing leaves it to Hadoop to decide.

_mroNumReduce :: Maybe Int

Number of reduce tasks; Nothing leaves it to Hadoop to decide.

_mroTaskTimeout :: Maybe Int

Timeout in miliseconds for Hadoop tasks. Set to 0 to disable the timeout.

_mroCompress :: Maybe String

Whether to use compression on reduce output.

_mroOutSep :: Maybe Char

Separator to be communicated to Hadoop for the reduce output. Sets both the 'stream.reduce.output.field.separator' and 'mapred.textoutputformat.separator' parameters. Sometimes useful to trick Hadoop into agreeing that the reduce output has both a key and a value, therefore avoiding the trailing separator forcefully inserted by Hadoop.

If you're outputting CSV for example, you may want to specify 'Just ,' here so that with 2 fields Hadoop will think you already have the key-value pair.

_mroReduceError :: ReduceErrorStrategy

What to do on reducer error.

data PartitionStrategy

Constructors

NoPartition
Expect a single key segment emitted from the Mapper.

Partition
Expect a composite key emitted from the Mapper.

Fields

keySegs :: Int

Total segments comprising a key.

partSegs :: Int

First N key segments used for partitioning. All keys that share these segments will be routed to the same reducer.

Low-level Utilities

mapper

Arguments

:: Mapper ByteString CompositeKey ByteString
A key/value producer - don't worry about putting any newline endings yourself, we do that for you.

-> IO ()

mapperWith :: Prism' ByteString t -> Mapper ByteString CompositeKey t -> IO ()

Construct a mapper program using given serialization Prism.

combiner :: MROptions -> Prism' ByteString b -> Reducer CompositeKey b (CompositeKey, b) -> IO ()

reducer

Arguments

:: MROptions

-> Prism' ByteString a
Input conversion function

-> Reducer CompositeKey a b
A step function for any given key. Will be rerun from scratch for each unique key based on MROptions.

-> Producer (ResourceT IO) b

Create a reducer program.

setLineBuffering :: MonadIO m => m ()

Data Serialization Utilities

module Hadron.Protocol