Hadron.Controller

Description

High level flow-control of Hadoop programs with ability to define a sequence of Map-Reduce operations in a Monad, have strongly typed data locations.

Hadoop Program Construction

data Controller a

All MapReduce steps are integrated in the Controller monad.

Warning: We do have an io combinator as an escape valve for you to use. However, you need to be careful how you use the result of an IO computation. Remember that the same main function will run on both the main orchestrator process and on each and every map/reduce node.

connect :: MapReduce a b -> [Tap a] -> Tap b -> Maybe String -> Controller ()

Connect a typed MapReduce program you supply with a list of sources and a destination.

connect'

Arguments

:: MapReduce a b
MapReduce step to run

-> [Tap a]
Input files

-> Protocol' b
Serialization protocol to be used on the output

-> Maybe String
A custom name for the job

-> Controller (Tap b)

Connect a MapReduce program to a set of inputs, returning the output tap that was implicity generated (on hdfs) in the process.

io :: IO a -> Controller a

Perform an IO operation both on the orchestrator and on the worker nodes.

orchIO :: IO a -> Controller ()

Perform an IO operation only on the orchestrator

nodeIO :: IO a -> Controller a

Perform an IO operation only on the worker nodes.

setVal :: String -> ByteString -> Controller ()

Set a persistent variable in Controller state. This variable will be set during main M-R job controller loop and communicated to all the map and reduce nodes and will be available there.

getVal :: String -> Controller (Maybe ByteString)

Get varible from Controller state

runOnce :: Serialize a => IO a -> Controller a

Perform an IO action in orchestrator to obtain value, then cache it on HDFS and magically make it available to nodes during their runtime.

data MapReduce a b

A packaged MapReduce step. Make one of these for each distinct map-reduce step in your overall Controller flow.

Constructors

forall k v . MRKey k => MapReduce

Fields

_mrOptions :: MROptions

Hadoop and MapReduce options affecting only this specific job.

_mrInPrism :: Prism' ByteString v

A serialization scheme for values between the map-reduce steps.

_mrMapper :: Mapper a k v

_mrCombiner :: Maybe (Reducer k v (k, v))

_mrReducer :: Either (Reducer k v b) (Conduit v (ResourceT IO) b)

Either a reducer or a final value converter for a map-only MapReduce job.

mrOptions :: Lens' (MapReduce a b) MROptions

type Mapper a k b = Conduit a (ResourceT IO) (k, b)

A Mapper parses and converts the unbounded incoming stream of input data into a stream of (key, value) pairs.

A mapper is responsible for signaling its own Nothing to signify it is done writing. Hadron will not automatically mark EOF on the OutputStream, which is basically stdout.

type Reducer k a r = Conduit (k, a) (ResourceT IO) r

A reducer takes an incoming stream of (key, value) pairs and emits zero or more output objects of type r.

Note that this framework guarantees your reducer function (i.e. the conduit you supply here) will see ONLY keys that are deemed equivalent based on the MROptions you supply. Different keys will be given to individual and isolated invocations of your reducer function. This is a very central abstraction (and one of the few major ones) provided by this framework.

It does not matter if you supply your own EOF signals via Nothing as Hadron will simply discard them before relaying over to stdout and supply its own EOF based on when its input is finished.

(>.>) :: MapReduce a b -> Conduit b (ResourceT IO) c -> MapReduce a c

Do something with m-r output before writing it to a tap.

(<.<) :: Conduit c (ResourceT IO) a -> MapReduce a b -> MapReduce c b

Do something with the m-r input before starting the map stage.

class MRKey k where

Methods

toCompKey :: k -> CompositeKey

keyParser :: Parser k

numKeys :: k -> Int

type CompositeKey = [Key]

newtype SingleKey a

Constructors

SingleKey

Fields

unKey :: a

newtype WrapSerialize a

Constructors

WrapSerialize

Fields

_getSerialized :: a

newtype WrapSafeCopy a

Constructors

WrapSafeCopy

Fields

_getSafeCopy :: a

Data Sources

data Tap a

Tap is a data source/sink definition that *knows* how to serve records of type a.

It comes with knowledge on how to decode ByteString to target type and can be used both as a sink (to save data form MR output) or source (to feed MR programs).

Usually, you just define the various data sources and destinations your MapReduce program is going to need:

customers = 'tap' "s3n://my-bucket/customers" (csvProtocol def)

Constructors

Tap

Fields

_tapLocation :: [FilePath]

_tapProto :: Protocol' a

tapProto :: forall a a. Lens (Tap a) (Tap a) (Protocol' a) (Protocol' a)

tapLocation :: forall a. Lens' (Tap a) [FilePath]

tap :: FilePath -> Protocol' a -> Tap a

Construct a DataDef

taps :: [FilePath] -> Protocol' a -> Tap a

mergeTaps :: Tap a -> Tap b -> Tap (Either a b)

Combine two taps intelligently into the Either sum type.

Matches on the prefix path given as part of each tap. It would therefore fail to work properly on self joins where the same data location is used in both taps.

concatTaps :: [Tap a] -> Tap a

binaryDirTap

Arguments

:: FilePath
A root location to list files under

-> (FilePath -> Bool)
A filter condition to refine the listing

-> Controller (Tap (FilePath, ByteString))

Creates a tap for a directory of binary files.

setupBinaryDir :: RunContext -> FilePath -> (FilePath -> Bool) -> IO (LocalFile, FilePath)

Grab list of files in destination, write into a file, put file on HDFS so it is shared and return the (local, hdfs) paths.

fileListTap

Arguments

:: RunContext

-> FilePath
A file containing a list of files to be used as input

-> Tap (FilePath, ByteString)

Tap for handling file lists. Hadoop can't process raw binary data because it splits on newlines. This tap allows you to get around that limitation by instead making your input a list of file paths that contain binary data. Then the file names get split by hadoop and each map job reads from those files as its first step.

fanOutTap

Arguments

:: RunContext

-> FilePath
Static location where fanout statistics will be written via the regular hadoop output.

-> FilePath
A temporary location where in-progress files can be kept.

-> (a -> FilePath)
Decision dispatch of where each object should go. Make sure to provide fully qualified hdfs directory paths; a unique token will be appended to each file based on the node producing it.

-> Conduit a (ResourceT IO) ByteString
How to serialize each object. Make sure this conduit provides for all the typical requirements: One record per line, no newlines inside the record, etc.

-> FanOutSink
How to sink the fanout, exposed here for flexibility.

-> Tap a

Sink objects into multiple output files through concurrent file-write processes behind the scenes. Work-around for Hadoop Streaming limitations in having to sink output into a single provided HDFS path.

sinkFanOut :: FanOutSink

Sink a stream into FanOut.

sequentialSinkFanout :: FanOutSink

A fanout that keeps only a single file open at a time. Each time the target filename changes, this will close/finalize the file and start the new file.

readTap :: RunContext -> Tap a -> IO [a]

Given a tap directory, enumerate and load all files inside. Caution: This is meant only as a way to load small files, or else you'll fill up your memory.

readHdfsFile :: RunContext -> Conduit ByteString (ResourceT IO) (FilePath, ByteString)

Conduit that takes in hdfs filenames and outputs the file contents. Will unpack .gz files automatically.

Command Line Entry Point

hadoopMain

Arguments

:: (MonadThrow m, MonadMask m, MonadIO m, Functor m, Applicative m)

=> [(AppLabel, Controller ())]
Hadoop streaming applications that can be run. First element of tuple is used to lookup the right application to run from the command line.

-> RunContext
Hadoop environment info.

-> RerunStrategy
What to do if destination files already exist.

-> m ()

The main entry point. Use this function to produce a command line program that encapsulates everything.

When run without arguments, the program will orchestrate the entire MapReduce job flow. The same program also doubles as the actual mapper/reducer executable when called with right arguments, though you don't have to worry about that.

data HadoopEnv

Constructors

HadoopEnv

Fields

_hsBin :: String

_hsJar :: String

Settings for MapReduce Jobs

data MROptions

Options for a single-step MR job.

mroPart :: Lens' MROptions PartitionStrategy

mroNumMap :: Lens' MROptions (Maybe Int)

mroNumReduce :: Lens' MROptions (Maybe Int)

mroCompress :: Lens' MROptions (Maybe String)

mroOutSep :: Lens' MROptions (Maybe Char)

mroTaskTimeout :: Lens' MROptions (Maybe Int)

data PartitionStrategy

Constructors

NoPartition
Expect a single key segment emitted from the Mapper.

Partition
Expect a composite key emitted from the Mapper.

Fields

keySegs :: Int

Total segments comprising a key.

partSegs :: Int

First N key segments used for partitioning. All keys that share these segments will be routed to the same reducer.

data Comparator

Constructors

RegularComp
Regular sorting

NumericComp Int Int Bool
Numeric sorting spanning fields i to j, True=reversed

data RerunStrategy

What to do when we notice that a destination file already exists.

Constructors

RSFail
Fail and log the problem.

RSReRun
Delete the file and rerun the analysis

RSSkip
Consider the analaysis already done and skip.

Hadoop Utilities

emitCounter

Arguments

:: ByteString
Group name

-> ByteString
Counter name

-> Integer
Increment

-> IO ()

Emit a counter to be captured, added up and reported by Hadoop.

hsEmitCounter :: ByteString -> Integer -> IO ()

Emit counter from this library's group

emitStatus :: ByteString -> IO ()

Emit a status line.

getFileName :: MonadIO m => m FilePath

Get the current filename from Hadoop ENV. Useful when writing Mappers and you would like to know what file you're currently dealing with.

MapReduce Combinators

mapReduce :: t

firstBy :: t

mapMR :: t

oneSnap :: t

joinMR

Arguments

:: (MRKey k, Monoid v, Serialize v)

=> Mapper (Either a b) k (Either v v)
Mapper for the input

-> MapReduce (Either a b) v

Monoidal inner (map-side) join for two types. Each type is mapped into the common monoid, which is then collapsed during reduce.

Make sure an incoming Left stays Left and a Right stays a Right.

TODO: Wrap around this with a better API so the user doesn't have to care.

joinStep

Arguments

:: (Show b, Monoid b, Serialize b, MRKey k)

=> [(Tap a, JoinType, Mapper a k b)]
Dataset definitions and how to map each dataset.

-> MapReduce a b

A convenient way to express map-sde multi-way join operations into a single data type. All you need to supply is the map operation for each tap, the reduce step is assumed to be the Monoidal mconcat.

joinMR is probably easier to use if you can get by with an inner join.

data JoinType

Constructors

JRequired

JOptional

type JoinKey = ByteString

Data Serialization Utilities

module Hadron.Protocol

module Hadron.Run