Hadron.Join

Documentation

type DataDefs = [(DataSet, JoinType)]

newtype DataSet

Constructors

DataSet

Fields

getDataSet :: ByteString

data JoinType

Constructors

JRequired

JOptional

type JoinKey = ByteString

joinMain

Arguments

:: (Serialize r, Monoid r, Show r)

=> DataDefs
Define your tables

-> (String -> DataSet)
Infer dataset from input filename

-> (DataSet -> Mapper ByteString CompositeKey r)
Map input stream to a join key and the common-denominator uniform data type we know how to mconcat.

-> Prism' ByteString r
Choose serialization method for final output.

-> IO ()

Make a stand-alone program that can act as a mapper and reducer, performing the join defined here.

For proper higher level operation, see the Controller module.

joinMapper

Arguments

:: (String -> DataSet)
Infer dataset from current filename

-> (DataSet -> Mapper a CompositeKey r)
Given a dataset, map it to a common data type

-> Mapper a CompositeKey r

Helper for easy construction of specialized join mapper.

This mapper identifies the active dataset from the currently streaming filename and uses filename to determine how the mapping shoudl be done.

joinReducer

Arguments

:: (Show r, Monoid r)

=> [(DataSet, JoinType)]
Table definitions

-> Reducer CompositeKey r r

Make join reducer from given table definitions

joinOpts :: MROptions

TODO: To be put into an Internal module

data JoinAcc a

We are either buffering input rows or have started streaming, as we think we're now receiving the last table we were expecting.

Constructors

Buffering

Fields

bufData :: !(HashMap DataSet [a])

Buffer of in-memory retained data. We have to retain (n-1) of the input datasets and we can then start emitting rows in constant-space for the nth dataset.

Streaming

Fields

strStems :: Vector a

bufToStr

Arguments

:: Monoid a

=> DataDefs
Table definitions for the current join

-> JoinAcc a
Buffering

-> JoinAcc a
Streaming

Convert a buffering state to a ready-to-stream state. Once this conversion is done, we'll start emitting output rows immediately and in constant space.