Hadron.Protocol

Documentation

data Protocol b a

A Protocols is a serialization strategy when we're dealing with streams of records, allowing for arbitrary data formats, delimiters and other potential obstructions.

Most of the time we'll be using 'Protocol\''s.

Constructors

Protocol

Fields

_protoEnc :: Conduit a (ResourceT IO) b

_protoDec :: Conduit b (ResourceT IO) a

type Protocol' a = Protocol ByteString a

Like Protocol but fixes the source format as ByteString.

protoEnc :: forall b a. Lens' (Protocol b a) (Conduit a (ResourceT IO) b)

protoDec :: forall b a. Lens' (Protocol b a) (Conduit b (ResourceT IO) a)

prismToProtocol :: Prism' ByteString a -> Protocol' a

Lift Prism to work with a newline-separated stream of objects.

It is assumed that the prism you supply to this function does not add newlines itself. You need to make them newline-free for this to work properly.

filterP :: (a -> Bool) -> Protocol b a -> Protocol b a

Filter elements in both directions of the protocol.

base64SerProtocol :: Serialize a => Protocol' a

Channel the Serialize instance through Base64 encoding to make it newline-safe, then turn into newline-separated stream.

base64SafeCopyProtocol :: SafeCopy a => Protocol' a

Channel the Serialize instance through Base64 encoding to make it newline-safe, then turn into newline-separated stream.

idProtocol :: Protocol' ByteString

Basically id from Control.Category. Just pass the incoming ByteString through.

linesProtocol :: Protocol' ByteString

A simple serialization strategy that works on lines of strings.

gzipProtocol :: Protocol ByteString ByteString

Encode and decode a gzip stream

showProtocol :: (Read a, Show a) => Protocol' a

Use 'Show'/'Read' instances to stream-serialize. You must be careful not to have any newline characters inside, or the stream will get confused.

This is meant for debugging more than anything. Do not use it in serious matters. Use serProtocol instead.

csvProtocol :: CSV b a => CSVSettings -> Protocol b a

Protocol for converting to/from any stream type b and CSV type a.

Serialization Prisms

pSerialize :: Serialize a => Prism' ByteString a

Serialize with the Serialize instance coupled with Base64 encoding, so it's free of restrictied characters.

Any Prism can be used as follows:

>>> import Control.Lens

To decode ByteString into target object:

>>> firstOf myPrism byteStr
Just a

To encode an object into ByteString:

>>> firstOf (re myPrism) myObject
Just byteStr

pSafeCopy :: SafeCopy a => Prism' ByteString a

pShow :: (Show a, Read a) => Prism' ByteString a

Serialize with the Show/Read instances

Serialization Utils

serialize :: Serialize a => a -> ByteString

Helper for reliable serialization through Base64 encoding so it is newline-free.

deserialize :: Serialize c => ByteString -> Either String c

Helper for reliable deserialization

Utils

lineC

Arguments

:: MonadThrow m

=> Int
Number of key segments (usually just 1), but may be higher if you're using multiple parts in your key.

-> Conduit ByteString m (CompositeKey, ByteString)

Parse lines of (key,value) for hadoop reduce stage

linesConduit :: MonadThrow m => Conduit ByteString m ByteString

Turn incoming stream into a stream of lines. This will automatically eat tab characters at the end of the line.

mkKey :: [ByteString] -> ByteString

Useful when making a key from multiple pieces of data

eitherPrism :: Prism' ByteString a -> Prism' ByteString b -> Prism' ByteString (Either a b)

Only works when a and b are disjoint types.

eitherProtocol :: Prism' ByteString a -> Prism' ByteString b -> Protocol' (Either a b)