Hadron.Run

Description

An operational run layer that either passes commands on to hadoop or runs things locally.

Documentation

data RunContext

Dispatch on the type of run

Constructors

LocalRun LocalRunSettings
Development mode: Emulate a hadoop run locally on this machine

HadoopRun HadoopEnv LocalRunSettings
Production mode: Actually run on hadoop. However, some utilites use local facilities so we still force you to have a local policy.

data LocalRunSettings

Constructors

LocalRunSettings

Fields

_lrsTempPath :: FilePath

Root of the "file system" during a localrun

data HadoopEnv

Constructors

HadoopEnv

Fields

_hsBin :: String

_hsJar :: String

clouderaDemo :: HadoopEnv

Settings for the cloudera demo VM.

amazonEMR :: HadoopEnv

Settings for Amazon's EMR instances.

data PartitionStrategy

Constructors

NoPartition
Expect a single key segment emitted from the Mapper.

Partition
Expect a composite key emitted from the Mapper.

Fields

keySegs :: Int

Total segments comprising a key.

partSegs :: Int

First N key segments used for partitioning. All keys that share these segments will be routed to the same reducer.

numSegs :: PartitionStrategy -> Int

Number of total key segments.

eqSegs :: PartitionStrategy -> Int

Number of key segments that constitute input object equality, for hadoop partitions.

data Comparator

Constructors

RegularComp
Regular sorting

NumericComp Int Int Bool
Numeric sorting spanning fields i to j, True=reversed

data HadoopRunOpts

Useful reference for hadoop flags:

http://hadoop.apache.org/docs/r2.4.1/hadoop-mapreduce-client/hadoop-mapreduce-client-core/mapred-default.xml

Constructors

HadoopRunOpts

Fields

_mrsInput :: [String]

_mrsOutput :: String

_mrsPart :: PartitionStrategy

_mrsNumMap :: Maybe Int

_mrsNumReduce :: Maybe Int

_mrsTaskTimeout :: Maybe Int

_mrsCombine :: Bool

_mrsCompress :: Maybe Codec

_mrsOutSep :: Maybe Char

A separator to be used in reduce output. It is sometimes useful to specify one to trick Hadoop.

_mrsJobName :: Maybe String

_mrsComparator :: Comparator

mrSettings

Arguments

:: [String]
Input files

-> String
Output files

-> HadoopRunOpts

A simple starting point to defining HadoopRunOpts

type Codec = String

gzipCodec :: Codec

snappyCodec :: Codec

launchMapReduce :: MonadIO m => RunContext -> String -> String -> HadoopRunOpts -> ExceptT String m ()

hdfsTempFilePath :: MonadIO m => RunContext -> FilePath -> m FilePath

Given a filename, produce an HDFS path for me in our temporary folder.

hdfsFileExists :: RunContext -> FilePath -> IO Bool

hdfsDeletePath :: RunContext -> FilePath -> IO ()

hdfsLs :: RunContext -> FilePath -> IO [File]

hdfsPut :: RunContext -> LocalFile -> FilePath -> IO ()

hdfsMkdir :: RunContext -> FilePath -> IO ()

hdfsCat :: RunContext -> FilePath -> Producer (ResourceT IO) ByteString

hdfsGet :: RunContext -> FilePath -> IO LocalFile

Copy a file from HDFS into local.

hdfsLocalStream :: RunContext -> FilePath -> Producer (ResourceT IO) ByteString

Copy a file down to local FS, then stream its content.

randomRemoteFile :: RunContext -> IO FilePath

randomLocalFile :: MonadIO m => m LocalFile

newtype LocalFile

Constructors

LocalFile

Fields

_unLocalFile :: FilePath

randomFileName :: MonadIO m => m LocalFile

withLocalFile :: MonadIO m => RunContext -> LocalFile -> (FilePath -> m b) -> m b

Helper to work with relative paths using Haskell functions like readFile and writeFile.

withRandomLocalFile :: MonadIO m => RunContext -> (FilePath -> m b) -> m LocalFile

module Hadron.Run.FanOut

hdfsFanOut

Arguments

:: RunContext

-> FilePath
A temporary folder where in-progress files can be placed

-> IO FanOut