Hadron.Run.FanOut

Description

Haskell-native ability to stream output to multiple files in Hadoop Streaming.

Documentation

data FanOut

Concurrent multi-output manager.

mkFanOut

Arguments

:: (FilePath -> IO (Handle, IO ()))
Open a handle for a given target path

-> IO FanOut

Make a new fanout manager that will use given process creator. Process is expected to pipe its stdin into the desired location.

fanWrite :: FanOut -> FilePath -> ByteString -> IO ()

Write into a file. A new process will be spawned if this is the first time writing into this file.

fanClose :: FanOut -> FilePath -> IO ()

Close a specific file.

fanCloseAll :: FanOut -> IO ()

Close all files. The same FanOut can be used after this, which would spawn new processes to write into files.

fhPath :: Lens' FileHandle FilePath

type FanOutSink = forall m a. MonadIO m => (a -> FilePath) -> (a -> m ByteString) -> FanOut -> Consumer a m Int

sinkFanOut :: FanOutSink

Sink a stream into FanOut.

sequentialSinkFanout :: FanOutSink

A fanout that keeps only a single file open at a time. Each time the target filename changes, this will close/finalize the file and start the new file.

fanStats :: FanOut -> IO (Map FilePath Int)

Grab # of writes into each file so far.

Additional lenses

chunkOrig :: Lens' FileChunk FilePath

chunkTarget :: Lens' FileChunk FilePath