Hadron.Run.Hadoop

Description

Deal with the hadoop command line program.

Documentation

data HadoopEnv

Constructors

HadoopEnv

Fields

_hsBin :: String

_hsJar :: String

hsBin :: Lens' HadoopEnv String

hsJar :: Lens' HadoopEnv String

clouderaDemo :: HadoopEnv

Settings for the cloudera demo VM.

amazonEMR :: HadoopEnv

Settings for Amazon's EMR instances.

data PartitionStrategy

Constructors

NoPartition
Expect a single key segment emitted from the Mapper.

Partition
Expect a composite key emitted from the Mapper.

Fields

keySegs :: Int

Total segments comprising a key.

partSegs :: Int

First N key segments used for partitioning. All keys that share these segments will be routed to the same reducer.

numSegs :: PartitionStrategy -> Int

Number of total key segments.

eqSegs :: PartitionStrategy -> Int

Number of key segments that constitute input object equality, for hadoop partitions.

data Comparator

Constructors

RegularComp
Regular sorting

NumericComp Int Int Bool
Numeric sorting spanning fields i to j, True=reversed

data HadoopRunOpts

Useful reference for hadoop flags:

http://hadoop.apache.org/docs/r2.4.1/hadoop-mapreduce-client/hadoop-mapreduce-client-core/mapred-default.xml

Constructors

HadoopRunOpts

Fields

_mrsInput :: [String]

_mrsOutput :: String

_mrsPart :: PartitionStrategy

_mrsNumMap :: Maybe Int

_mrsNumReduce :: Maybe Int

_mrsTaskTimeout :: Maybe Int

_mrsCombine :: Bool

_mrsCompress :: Maybe Codec

_mrsOutSep :: Maybe Char

A separator to be used in reduce output. It is sometimes useful to specify one to trick Hadoop.

_mrsJobName :: Maybe String

_mrsComparator :: Comparator

mrSettings

Arguments

:: [String]
Input files

-> String
Output files

-> HadoopRunOpts

A simple starting point to defining HadoopRunOpts

mrsInput :: Lens' HadoopRunOpts [String]

mrsOutput :: Lens' HadoopRunOpts String

mrsPart :: Lens' HadoopRunOpts PartitionStrategy

mrsNumMap :: Lens' HadoopRunOpts (Maybe Int)

mrsNumReduce :: Lens' HadoopRunOpts (Maybe Int)

mrsCombine :: Lens' HadoopRunOpts Bool

mrsCompress :: Lens' HadoopRunOpts (Maybe Codec)

mrsOutSep :: Lens' HadoopRunOpts (Maybe Char)

mrsJobName :: Lens' HadoopRunOpts (Maybe String)

mrsComparator :: Lens' HadoopRunOpts Comparator

mrsTaskTimeout :: Lens' HadoopRunOpts (Maybe Int)

type Codec = String

gzipCodec :: Codec

snappyCodec :: Codec

Hadoop Command Line Wrappers

hadoopMapReduce :: MonadIO m => HadoopEnv -> MapReduceKey -> RunToken -> HadoopRunOpts -> ExceptT String m ()

hdfsFileExists :: HadoopEnv -> FilePath -> IO Bool

Check if the target file is present.

hdfsDeletePath :: HadoopEnv -> FilePath -> IO ()

Check if the target file is present.

hdfsLs :: HadoopEnv -> FilePath -> IO [File]

List a directory's contents

parseLS :: String -> String -> [File]

TODO: The lcs function does not guarantee contiguous-common regions, so this function may behave strangely. We should figure out a way to use longest-common-prefix like semantics.

hdfsPut :: HadoopEnv -> FilePath -> FilePath -> IO ()

Copy file from a location to a location

hdfsFanOut :: HadoopEnv -> FilePath -> IO FanOut

hdfsFanOutStream :: HadoopEnv -> FilePath -> IO FanOut

Create a new multiple output file manager.

hdfsMkdir :: HadoopEnv -> String -> IO ()

Create HDFS directory if missing

tmpRoot :: FilePath

hdfsCat :: MonadIO m => HadoopEnv -> FilePath -> Producer m ByteString

Stream data directly from HDFS using hdfs cat.

NOTE: It appears that this function may output a header before the file contents. Be careful!

hdfsGet :: HadoopEnv -> FilePath -> FilePath -> IO ()

Copy file from HDFS to a temporary local file whose name is returned.

hdfsChmod

Arguments

:: HadoopEnv

-> String
Target path

-> String
Permissions string

-> IO ExitCode

Apply recursive permissions to given

randomFilename :: HadoopEnv -> IO FilePath

Generates a random filename in the tmphadron directory.