Hadron.Run.Local

Description

Emulate all hadoop operations locally

Documentation

newtype LocalFile

Constructors

LocalFile

Fields

_unLocalFile :: FilePath

unLocalFile :: Iso' LocalFile FilePath

data LocalRunSettings

Constructors

LocalRunSettings

Fields

_lrsTempPath :: FilePath

Root of the "file system" during a localrun

lrsTempPath :: Iso' LocalRunSettings FilePath

type Local = ReaderT LocalRunSettings IO

runLocal :: r -> ReaderT r m a -> m a

path :: (MonadIO m, MonadReader LocalRunSettings m) => LocalFile -> m FilePath

getRecursiveDirectoryContents :: FilePath -> IO [FilePath]

getDirectoryContents' :: FilePath -> IO [FilePath]

A version that return [] instead of an error when directory does not exit.

getInputFiles :: FilePath -> IO [FilePath]

Recursive contents if given a directory, assumed glob pattern if not.

localMapReduce

Arguments

:: MonadIO m

=> LocalRunSettings

-> String
MapReduceKey

-> String
RunToken

-> HadoopRunOpts

-> ExceptT String m ()

echo :: (Applicative m, MonadIO m) => Severity -> LogStr -> m ()

echoInfo :: (Applicative m, MonadIO m) => LogStr -> m ()

clearExit :: MonadIO m => ExceptT String m ExitCode -> ExceptT [Char] m ()

Fail if command not successful.

hdfsFileExists :: (MonadIO m, MonadReader LocalRunSettings m) => LocalFile -> m Bool

Check if the target file is present.

hdfsDeletePath :: (MonadIO m, MonadReader LocalRunSettings m) => LocalFile -> m ()

hdfsLs :: (MonadIO m, MonadReader LocalRunSettings m) => LocalFile -> m [File]

hdfsPut :: (MonadIO m, MonadReader LocalRunSettings m) => LocalFile -> LocalFile -> m ()

hdfsFanOut

Arguments

:: (MonadIO m, MonadReader LocalRunSettings m)

=> FilePath
Temporary file location.

-> m FanOut

Create a new multiple output file manager.

hdfsMkdir :: (MonadIO m, MonadReader LocalRunSettings m) => LocalFile -> m ()

hdfsCat :: LocalFile -> Producer (ResourceT Local) ByteString

hdfsGet :: (MonadIO m, MonadReader LocalRunSettings m) => LocalFile -> m LocalFile

hdfsLocalStream :: LocalFile -> Producer (ResourceT Local) ByteString

randomFileName :: MonadIO m => m LocalFile

withLocalFile

Arguments

:: MonadIO m

=> LocalRunSettings

-> LocalFile
A relative path in our working folder

-> (FilePath -> m b)
What to do with the absolute path.

-> m b

Helper to work with relative paths using Haskell functions like readFile and writeFile.