Hadron.Utils

Documentation

performEvery

Arguments

:: Monad m

=> Integer
Once every N items that flow throught he pipe.

-> (Integer -> m ())

-> ConduitM a a m ()

Perform a given monadic action once every X elements flowing through this conduit.

data File

Constructors

File

Fields

_filePerms :: String

_fileSize :: Int

_fileDate :: String

_fileTime :: String

_filePath :: String

fileTime :: Lens' File String

fileSize :: Lens' File Int

filePerms :: Lens' File String

filePath :: Lens' File String

fileDate :: Lens' File String

parseLs :: String -> Maybe File

randomToken :: Int -> IO String

Generate a random token

(!?) :: [a] -> Int -> Maybe a